Staging (#30)
* devops Initial Commit of CI/CD pipeline

* devops

* devops Fix build hook

* devops Fix build hook v2

* devops Fix build hook v3

* devops Fix build hook v4

* devops Fix build hook v5

* devops Fix build hook v5.1

* devops Add s3 storage

* devops Setup the CI/CD properly for staging & production

* devops Setup the CI/CD properly for staging & production v2

* sync vapor stuff. complete seeders. (#24)

* sync vapor stuff. complete seeders.

* cleanup

* sync vapor stuff. complete seeders.

* image handling cleanup

* staging Enable installation of dev packages in staging (needed for seeders)

* resolve VotingPower seeder broken

* staging Install Vapor-UI to monitor the app

* staging Implement ability to run lucid container service in standalone mode

* staging Fix port issue on standalone lucid container

* disable registration when voting power is missing (#25)

* register ballot tlc (#26)

* resolve snapshot import failing on large files. Ballot questions cleanup. (#27)

* ui refinements. support multi questions on ballot (#28)

* admin ballot cleanup. wait for onchain confirmation. public ui cleanup (#29)

* working on petition and poll modules (#31)

---------

+++ application/app/Http/Controllers/Admin/BallotController.php
@@ -29,15 +29,59 @@ use Inertia\Response;
 use JetBrains\PhpStorm\NoReturn;
 use Momentum\Modal\Modal;
 use Saloon\Exceptions\Request\FatalRequestException;
+use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Hash;
 
 class BallotController extends Controller
 {
+    public function index()
+    {
+        $ballots = Ballot::all();
+        $crumbs = [
+
+            [
+                'label' => 'Ballots',
+                'link' => route('admin.ballots.index')
+            ],
+        ];
+
+        return Inertia::render(
+            'Auth/Ballot/Index',
+            [
+                'ballots' => $ballots,
+                'crumbs' => $crumbs,
+            ]
+        );
+    }
+
+    public function ballotsData(Request $request)
+    {
+        $page = $request->query('page') ?? 1;
+        $perPage = $request->query('perPage') ?? 6;
+
+        $ballots = Ballot::paginate($perPage, ['*'], 'page', $page);
+
+        return BallotData::collection($ballots);
+    }
+
     /**
      * Display the new ballot's form.
      */
     public function create(Request $request): Response
     {
-        return Inertia::render('Auth/Ballot/Create', []);
+        $crumbs = [
+            [
+                'label' => 'Ballots',
+                'link' => route('admin.ballots.index')
+            ],
+            [
+                'label' => 'Create Ballot',
+                'link' => route('admin.ballots.create')
+            ],
+        ];
+        return Inertia::render('Auth/Ballot/Create', [
+            'crumbs' => $crumbs,
+        ]);
     }
 
     /**
@@ -45,8 +89,19 @@ class BallotController extends Controller
      */
     public function view(Request $request, Ballot $ballot): Response
     {
+        $crumbs = [
+            [
+                'label' => 'Ballots',
+                'link' => route('admin.ballots.index')
+            ],
+            [
+                'label' => 'View Ballot',
+                'link' => route('admin.ballots.view', ['ballot' => $ballot])
+            ],
+        ];
         return Inertia::render('Auth/Ballot/View', [
             'ballot' => BallotData::from($ballot->load('snapshot', 'questions.choices')),
+            'crumbs' => $crumbs,
         ]);
     }
 
@@ -57,9 +112,20 @@ class BallotController extends Controller
     {
         $ballot->load(['questions.choices', 'snapshot', 'policies']);
 
+        $crumbs = [
+            [
+                'label' => 'Ballots',
+                'link' => route('admin.ballots.index')
+            ],
+            [
+                'label' => 'Edit Ballot',
+                'link' => route('admin.ballots.edit', ['ballot' => $ballot])
+            ],
+        ];
         return Inertia::render('Auth/Ballot/Edit', [
             'ballot' => BallotData::from($ballot),
-            'addresses' => $this->policyAddress($ballot),
+            'addresses' => $this->policyAddresses($ballot),
+            'crumbs' => $crumbs,
         ]);
     }
 
@@ -94,7 +160,6 @@ class BallotController extends Controller
             $ballot->title = $ballotData->title;
             $ballot->description = $ballotData->description;
             $ballot->version = $ballotData->version;
-            $ballot->status = $ballotData->status;
             $ballot->type = $ballotData->type;
             $ballot->started_at = $ballotData->started_at;
             $ballot->ended_at = $ballotData->ended_at;
@@ -109,19 +174,46 @@ class BallotController extends Controller
     /**
      * Delete the ballot's account.
      */
-    public function destroy(Request $request, $ballot): RedirectResponse
+    public function destroy(Request $request, Ballot $ballot): RedirectResponse
     {
-        $response = Gate::inspect('delete', $ballot);
+        $password = $request->input('password');
+        if(!$password){
+            return Redirect::back()->withErrors(['password' => 'Password is required.']);
+        }
 
-        if ($response->allowed()) {
-            $existingBallot = Ballot::byHash($ballot);
-            $existingBallot->delete();
+        if (Hash::check($password, Auth::user()->password)) {
+            $response = Gate::inspect('delete', $ballot);
 
-            return Redirect::to('/');
+            if ($response->allowed()) {
+                $existingBallot = Ballot::byHash($ballot->hash);
+                $existingBallot->delete();
+
+                return Redirect::route('admin.dashboard');
+            } else {
+                return Redirect::route('admin.ballots.view', ['ballot' => $ballot]);
+            }
         } else {
-            return Redirect::route('admin.ballots.view', ['ballot' => $ballot]);
+            return Redirect::back()->withErrors(['password' => 'Invalid password for current user.']);
         }
     }
+
+    /**
+     * Update the ballot's status.
+     */
+    public function statusUpdate(Ballot $ballot)
+    {
+        $ballot = Ballot::byHash($ballot->hash);
+        $response = $ballot->status == 'published' ? Gate::inspect('publish', $ballot) : Gate::inspect('update', $ballot);
+
+        if ($response->allowed()) {
+            $ballot->status = ModelStatusEnum::PUBLISHED;
+            $ballot->update();
+            return Redirect::back();
+        } else {
+            return Redirect::back()->withErrors(['error' => 'Not authorized']);
+        }
+    }
+
 
     /**
      * Create a ballot's question.
@@ -133,11 +225,15 @@ class BallotController extends Controller
     {
         $response = Gate::inspect('create', Question::class);
         if ($response->allowed()) {
+            $crumbs = [
+                ['label' => 'Create Question', 'link' => route('admin.ballots.questions.create', ['ballot' => $ballot->hash])],
+            ];
             return Inertia::modal('Auth/Question/Create')
                 ->with([
                     'ballot' => BallotData::from($ballot),
                     'questionTypes' => QuestionTypeEnum::values(),
                     'questionsStatuses' => ModelStatusEnum::values(),
+                    'crumbs' => $crumbs,
                 ])
                 ->baseRoute('admin.ballots.edit', [
                     'ballot' => $ballot->hash,
@@ -156,10 +252,14 @@ class BallotController extends Controller
     public function editQuestion(Request $request, Ballot $ballot, Question $question)
     {
         $ballot->load(['questions']);
+        $crumbs = [
+            ['label' => 'Edit Question', 'link' => route('admin.ballots.questions.edit', ['ballot' => $ballot->hash])],
+        ];
 
         return Inertia::render('Auth/Question/Edit', [
             'ballot' => BallotData::from($ballot),
             'question' => QuestionData::from($question),
+            'crumbs' => $crumbs,
         ]);
     }
 
@@ -283,7 +383,6 @@ class BallotController extends Controller
             return Redirect::back()->withErrors(['error' => $e->getMessage()]);
         }
     }
-
 
     /**
      * Store a newly created Ballot in storage.
@@ -415,7 +514,7 @@ class BallotController extends Controller
         return Redirect::back();
     }
 
-    public function policyAddress($ballot)
+    public function policyAddresses($ballot)
     {
         $firstPolicySeed = null;
         $secondPolicySeed = null;
@@ -443,7 +542,7 @@ class BallotController extends Controller
             $lucid = new LucidConnector;
             $policyResponse = $lucid->send($policyAddress);
 
-            $registrationPolicyAddress = $policyResponse->json()['address'];
+            $registrationPolicyAddress = $policyResponse->json()['address'] ?? null;
         }
 
         if ($secondPolicySeed) {
@@ -463,15 +562,33 @@ class BallotController extends Controller
         ];
     }
 
-    public function addImageLink(Request $request, Ballot $ballot)
+    public function addImageLink(Request $request, Ballot $ballot, Policy $policy)
     {
-        $registrationPolicy = $ballot->registration_policy()->first();
-        $registrationPolicy->image_link = $request->link;
-        $registrationPolicy->save();
-        return true;
+        $response = Gate::inspect('update', $ballot);
+        if ($response->allowed()) {
+            $policy->image_link = $request->link;
+            $policy->save();
+            return true;
+        } else {
+            $this->unauthenticated($request);
+        }
+        return false;
     }
 
-    public function destroyPolicy(Request $request, Ballot $ballot)
+    public function unLinkSnapShot(Ballot $ballot, Snapshot $snapshot)
     {
+        $currentDate = Carbon::now();
+        $startDate = Carbon::parse($ballot->started_at);
+
+        if ($currentDate->greaterThanOrEqualTo($startDate)) {
+            throw new \Exception ('Cannot remove snapshot, the ballot has already started!');
+        }
+
+        try {
+            $snapshot->ballot_id = null;
+            $snapshot->save();
+        } catch (\Exception $e) {
+            throw $e;
+        }
     }
 }
